MNT/FIX: Cleanup resources and fix issues during deployment
This is an update based on test deployments that fixes a few issues
that came up and adds comments about what was needed.
