+++ tests/infrastructure/test_indexer_lambda_construct.py
@@ -2,7 +2,6 @@
 
 import pytest
 from aws_cdk import aws_ec2 as ec2
-from aws_cdk import aws_lambda as lambda_
 from aws_cdk import aws_rds as rds
 from aws_cdk.assertions import Template
 
@@ -14,7 +13,7 @@ from sds_data_manager.constructs.networking_construct import NetworkingConstruct
 
 
 @pytest.fixture()
-def template(stack, env):
+def template(stack, env, code):
     """Indexer lambda setup."""
     data_bucket = DataBucketConstruct(stack, "indexer-data-bucket", env=env)
     networking_construct = NetworkingConstruct(stack, "Networking")
@@ -32,6 +31,8 @@ def template(stack, env):
         username="imap",
         secret_name="sdp-database-creds-rds",  # noqa
         database_name="imapdb",
+        code=code,
+        layers=[],
     )
     monitoring_construct = MonitoringConstruct(
         stack, construct_id="MonitoringConstruct"
@@ -39,7 +40,7 @@ def template(stack, env):
     IndexerLambda(
         stack,
         "indexer-lambda",
-        code=lambda_.Code.from_inline("def handler(event, context):\n    pass"),
+        code=code,
         db_secret_name="test-secrets",  # noqa
         vpc=networking_construct.vpc,
         vpc_subnets=database_construct.rds_subnet_selection,
@@ -55,6 +56,6 @@ def template(stack, env):
 
 def test_indexer_role(template):
     """Ensure the template has appropriate IAM roles."""
-    template.resource_count_is("AWS::IAM::Role", 4)
-    # 2 for RDS + 1 for indexer lambda
-    template.resource_count_is("AWS::Lambda::Function", 3)
+    template.resource_count_is("AWS::IAM::Role", 6)
+    # 4 for RDS stack + 1 for indexer lambda
+    template.resource_count_is("AWS::Lambda::Function", 5)
